fixed single value signals
- now arrays with only one value return a scalar instead of array
